+++ src/vector/blas/xpay.cpp
@@ -13,7 +13,7 @@ namespace dd_avx{
 
 #pragma omp parallel private(regs)
 		{
-			size_t i=0, is=0, ie=0;
+			int i=0, is=0, ie=0;
 			get_isie(y.size(), is, ie);
 			SIMDreg alpha_hi = broadcast(alpha.x[0]);
 			SIMDreg alpha_lo = broadcast(alpha.x[1]);
@@ -45,7 +45,7 @@ namespace dd_avx{
 
 #pragma omp parallel private(regs)
 		{
-			size_t i=0, is=0, ie=0;
+			int i=0, is=0, ie=0;
 			get_isie(y.size(), is, ie);
 			SIMDreg alpha_hi = broadcast(alpha.x[0]);
 			SIMDreg alpha_lo = broadcast(alpha.x[1]);
@@ -77,7 +77,7 @@ namespace dd_avx{
 
 #pragma omp parallel private(regs)
 		{
-			size_t i=0, is=0, ie=0;
+			int i=0, is=0, ie=0;
 			get_isie(y.size(), is, ie);
 			SIMDreg alpha_hi = broadcast(alpha.x[0]);
 			SIMDreg alpha_lo = broadcast(alpha.x[1]);
@@ -108,7 +108,7 @@ namespace dd_avx{
 
 #pragma omp parallel private(regs)
 		{
-			size_t i=0, is=0, ie=0;
+			int i=0, is=0, ie=0;
 			get_isie(y.size(), is, ie);
 			SIMDreg alpha_hi = broadcast(alpha.x[0]);
 			SIMDreg alpha_lo = broadcast(alpha.x[1]);
@@ -140,7 +140,7 @@ namespace dd_avx{
 
 #pragma omp parallel private(regs)
 		{
-			size_t i=0, is=0, ie=0;
+			int i=0, is=0, ie=0;
 			get_isie(y.size(), is, ie);
 			SIMDreg alpha_hi = broadcast(alpha);
 			SIMDreg alpha_lo = regs.zeros;
@@ -172,7 +172,7 @@ namespace dd_avx{
 
 #pragma omp parallel private(regs)
 		{
-			size_t i=0, is=0, ie=0;
+			int i=0, is=0, ie=0;
 			get_isie(y.size(), is, ie);
 			SIMDreg alpha_hi = broadcast(alpha);
 			SIMDreg alpha_lo = regs.zeros;
@@ -204,7 +204,7 @@ namespace dd_avx{
 
 #pragma omp parallel private(regs)
 		{
-			size_t i=0, is=0, ie=0;
+			int i=0, is=0, ie=0;
 			get_isie(y.size(), is, ie);
 			SIMDreg alpha_hi = broadcast(alpha);
 			SIMDreg alpha_lo = regs.zeros;
@@ -235,7 +235,7 @@ namespace dd_avx{
 
 #pragma omp parallel private(regs)
 		{
-			size_t i=0, is=0, ie=0;
+			int i=0, is=0, ie=0;
 			get_isie(y.size(), is, ie);
 			SIMDreg alpha_hi = broadcast(alpha);
 			SIMDreg alpha_lo = regs.zeros;
